pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scicoqa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv==2.2.0",
    "datasets>=4.1.1",
    "evaluate>=0.4.6",
    "gitpython>=3.1.45",
    "google-genai>=1.34.0",
    "joblib>=1.5.1",
    "jupyter>=1.1.1",
    "litellm>=1.76.0",
    "mistralai>=1.9.7",
    "nbconvert>=7.16.6",
    "openai>=1.107.0",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pygithub>=2.7.0",
    "pystemmer>=3.0.0",
    "python-json-logger>=3.3.0",
    "requests>=2.32.5",
    "rouge-score>=0.1.2",
    "sentence-transformers>=5.1.0",
    "simple-parsing>=0.1.7",
    "statsmodels>=0.14.6",
    "tinydb>=4.8.2",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "seaborn>=0.13.2",
    "huggingface-hub>=0.36.0",
    "openpyxl>=3.1.5",
]

[project.scripts]
scicoqa-batch = "scicoqa.core.batch:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scicoqa*"]
exclude = ["out*", "data*", "config*"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.10",
    "ipdb>=0.13.13",
]

[tool.ruff]
src = ["."]
extend-exclude = ["scicoqa/github/issue_img_to_text.py"]

line-length = 88

[tool.ruff.lint]
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes
select = ["E", "F", "I"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = false
order-by-type = true
from-first = false

section-order = [
    "future",
    "standard-library", 
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
